Engine: Addressing comments and moving taxonomix to concerns folder.
