--- app/models/katello/concerns/taxonomix.rb
@@ -1,68 +1,0 @@
-module Katello
-  module Concerns
-    module Taxonomix
-      extend ActiveSupport::Concern
-
-      included do
-        taxonomy_join_table = "taxable_taxonomies"
-        has_many taxonomy_join_table, :dependent => :destroy, :as => :taxable
-        has_many :locations,     :through => taxonomy_join_table, :source => :taxonomy,
-                 :conditions => "taxonomies.type='Location'", :validate => false
-        belongs_to :organization, :conditions => "taxonomies.type='Organization'", :validate => true
-        after_initialize :set_current_taxonomy
-
-        scoped_search :in => :locations, :on => :name, :rename => :location, :complete_value => true
-        scoped_search :in => :organization, :on => :name, :rename => :organization, :complete_value => true
-      end
-
-      module ClassMethods
-        def with_taxonomy_scope
-          scope =  block_given? ? yield : where('1=1')
-          scope = scope.where("#{self.table_name}.id in (#{inner_select(Location.current)}) #{user_conditions}") if SETTINGS[:locations_enabled] && Location.current && !Location.ignore?(self.to_s)
-          scope = scope.where(:organization_id => Organization.current.id) if SETTINGS[:organizations_enabled] and Organization.current && !Organization.ignore?(self.to_s)
-          scope.readonly(false)
-        end
-
-        def inner_select taxonomy
-          taxonomy_ids = Array.wrap(taxonomy).map(&:id).join(',')
-          "SELECT taxable_id from taxable_taxonomies WHERE taxable_type = '#{self.name}' AND taxonomy_id in (#{taxonomy_ids}) "
-        end
-
-        # I the case of users we want the taxonomy scope to get both the users of the taxonomy and admins.
-        # This is done here and not in the user class because scopes cannot be chained with OR condition.
-        def user_conditions
-          sanitize_sql_for_conditions([" OR users.admin = ?", true]) if self == User
-        end
-      end
-
-      def set_current_taxonomy
-        if self.new_record? && self.errors.empty?
-          self.locations    << Location.current     if add_current_location?
-          self.organization << Organization.current if add_current_organization?
-        end
-      end
-
-      def add_current_organization?
-        add_current_taxonomy?(:organization)
-      end
-
-      def add_current_location?
-        add_current_taxonomy?(:location)
-      end
-
-      def add_current_taxonomy?(taxonomy)
-        klass, association = case taxonomy
-                               when :organization
-                                 [Organization, :organization]
-                               when :location
-                                 [Location, :locations]
-                               else
-                                 raise ArgumentError, "unknown taxonomy #{taxonomy}"
-                             end
-        current_taxonomy = klass.current
-        ::Taxonomy.enabled?(taxonomy) && current_taxonomy && (!self.send(association) == current_taxonomy)
-      end
-
-    end
-  end
-end
